+++ premier-design/shared/ui/copyrighting/Copyrighting.tsx
@@ -1,24 +1,23 @@
-import React, {FC, ReactElement} from "react";
-import styles from "./Copyrighting.module.css";
-import {SITE_OPERATOR} from "@shared/constants/company";
-import Logo from "@shared/ui/logo/Logo";
-import SocialIcons from "@shared/ui/social-icons/SocialIcons";
-import {format} from "date-fns";
+import {type FC, type ReactElement} from 'react';
+import {format} from 'date-fns';
+import {SITE_OPERATOR} from '@shared/constants/company';
+import Logo from '@shared/ui/logo/Logo';
+import SocialIcons from '@shared/ui/social-icons/SocialIcons';
+import styles from './Copyrighting.module.css';
 
 const Copyrighting: FC = (): ReactElement => {
     return (
-
         <section className={styles.footer__contacts}>
             <div className={styles.footer__logo}>
-                <Logo/>
+                <Logo />
+            </div>
+            <div className={styles.footer__socialIcons}>
+                <SocialIcons />
             </div>
             <div className={styles.footer__copy}>
                 &copy; {SITE_OPERATOR.brandName}, {SITE_OPERATOR.copyrightStartYear} - {format(new Date(), 'yyyy')}.
-                <br/>
+                <br />
                 Все права защищены
-            </div>
-            <div className={styles.footer__socialIcons}>
-                <SocialIcons/>
             </div>
         </section>
     );