+++ premier-design/shared/ui/legal-requisites/index.ts
@@ -1,0 +1,1 @@
+export {default as LegalRequisites} from './LegalRequisites';