+++ premier-design/shared/ui/phone/Phone.tsx
@@ -1,25 +1,26 @@
+import {type ReactElement} from 'react';
 import Link from 'next/link';
 import {FaPhoneVolume} from 'react-icons/fa';
+import {SITE_OPERATOR} from '@shared/constants/company';
 import styles from './Phone.module.css';
-import {ReactElement} from "react";
 
 const Phone = (): ReactElement => {
     return (
         <div className={styles.phone}>
             <Link
-                href="tel:+375291942881"
+                href={`tel:${SITE_OPERATOR.phone.tel}`}
                 className={styles.phone__container}
                 aria-label="Позвонить"
             >
                 <div className={styles.phone__logo}>
-                    <FaPhoneVolume/>
+                    <FaPhoneVolume />
                 </div>
                 <div className={styles.phone__number}>
-                    <p>+375(29)194-28-81</p>
+                    <p>{SITE_OPERATOR.phone.display}</p>
                 </div>
             </Link>
         </div>
     );
-}
+};
 
 export default Phone;